requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
macaddress==2.0.2
Mako==1.2.3
MarkupSafe==2.1.1
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
pycryptodome==3.20.0
python-dateutil==2.8.2
pytz==2022.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.28.1
six==1.16.0
# get rid of the below
tplinkrouterc6u==3.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
